Resolved a warning in tests/src/test_SeedIndex.cpp.

+++ tests/src/test_SeedIndex.cpp
@@ -594,7 +594,7 @@ TEST(SeedIndex, CollectHitsLongSeedSpan)
         std::ostringstream oss;
         oss << result;
         SCOPED_TRACE("Testing hit: " + oss.str());
-        EXPECT_TRUE(static_cast<int32_t>(result.querySpan) >= static_cast<int32_t>(0));
+        EXPECT_TRUE(static_cast<int32_t>(result.querySpan) >= 0);
     }
 }
